Add an option to keep the data to its defaults.
I keep the blog data to its initial value when the server doesn't return
that specified value. This is preferable to keeping it undefined.

When the value changes to undefined a react error will pop up, so this
is done to avoid the error. The error is about controlled inputs
changing to uncontrolled.